requirements.txt
yt-dlp>=2024.12.13
google-generativeai>=0.7.2
python-dotenv>=1.0.0
langdetect>=1.0.9
